+++ src/components/Dropdown/index.js
@@ -2,110 +2,99 @@ import React from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { Manager, Popper, Target } from 'react-popper';
+import { withToggle } from '../../utils/recompose-utils';
 
-class Dropdown extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isOpen: false,
-      overChildren: false,
-    };
-    this.handleToggle = this.handleToggle.bind(this);
-    this.handleOnBlur = this.handleOnBlur.bind(this);
-    this.handleMouseOverChildren = this.handleMouseOverChildren.bind(this);
-    this.handleMouseLeavingChildren = this.handleMouseLeavingChildren.bind(this);
-  }
+const Dropdown = withToggle(({
+  arrowIcon,
+  buttonContent,
+  children,
+  hide,
+  isDisabled = false,
+  isOpen,
+  fullWidth,
+  placement = 'bottom-start',
+  show,
+  style,
+  testSection,
+  width = 200,
+  zIndex = 999,
+  toggle,
+}) => {
 
-  handleMouseOverChildren() {
-    this.setState({
-      overChildren: true,
-    });
-  }
+  const groupClass = classNames(
+    'oui-dropdown-group',
+    { ['width--1-1']: fullWidth }
+  );
 
-  handleMouseLeavingChildren() {
-    this.setState({
-      overChildren: false,
-    });
-  }
+  const buttonClass = classNames(
+    'oui-button',
+    { [`oui-button--${style}`]: style },
+    { ['oui-button--full soft--left text--left']: fullWidth }
+  );
 
-  handleOnBlur() {
-    if (!this.state.overChildren) {
-      this.setState({
-        isOpen: false,
-      });
+  const iconClass = classNames(
+    'push-half--left',
+    {
+      ['oui-arrow-inline--down']: arrowIcon === true || arrowIcon === 'down',
+      ['oui-arrow-inline--left']: arrowIcon === 'left',
+      ['oui-arrow-inline--right']: arrowIcon === 'right',
+      ['oui-arrow-inline--up']: arrowIcon === 'up',
     }
-  }
+  );
 
-  handleToggle(event) {
-    this.setState({
-      isOpen: !this.state.isOpen,
-      overChildren: false,
-    });
-  }
-
-  render() {
-    const {
-      buttonContent,
-      children,
-      isDisabled = false,
-      fullWidth,
-      placement = 'bottom-start',
-      style,
-      testSection,
-      width = 200,
-      zIndex = 999,
-    } = this.props;
-
-    const groupClass = classNames(
-      'oui-dropdown-group',
-      { ['width--1-1']: fullWidth }
-    );
-
-    const buttonClass = classNames(
-      'oui-button',
-      { [`oui-button--${style}`]: style },
-      { ['oui-button--full soft--left text--left']: fullWidth }
-    );
-
-    return (
-      <Manager
-        data-ui-component={ true }
-        className={ groupClass }
-        data-test-section={ testSection }>
-        <Target>
-          <button
-            type='button'
-            className={ buttonClass }
-            disabled={ isDisabled }
-            onClick={ this.handleToggle }
-            onBlur={ this.handleOnBlur }>
-            <div className='flex'>
-              <div className='flex--1 truncate'>{ buttonContent }</div>
-            </div>
-          </button>
-        </Target>
-        <Popper
-          placement={ placement }
-          className='oui-dropdown-children'
-          style={{
-            zIndex: zIndex,
-            position: 'absolute',
-            width: width,
-            marginTop: 2,
-            marginBottom: 2,
-            boxShadow: '0 2px 3px rgba(0,0,0,.1)',
-          }}
-          onMouseOver={ this.handleMouseOverChildren }
-          onMouseLeave={ this.handleMouseLeavingChildren }
-          onClick={ this.handleToggle } >
-          { this.state.isOpen && !isDisabled && children }
-        </Popper>
-      </Manager>
-    );
-  }
-}
+  return (
+    <Manager
+      data-ui-component={ true }
+      className={ groupClass }
+      data-test-section={ testSection }>
+      <Target>
+        <button
+          type='button'
+          className={ buttonClass }
+          disabled={ isDisabled }
+          onClick={ toggle }
+          onBlur={ hide }>
+          <div className='flex'>
+            <div className='flex--1 truncate'>{ buttonContent }</div>
+            {
+              !!arrowIcon && arrowIcon !== 'none' && (
+                <div className='text--right'><span className={ iconClass }></span></div>
+              )
+            }
+          </div>
+        </button>
+      </Target>
+      <Popper
+        placement={ placement }
+        className='oui-dropdown-children'
+        style={{
+          zIndex: zIndex,
+          position: 'absolute',
+          width: width,
+          marginTop: 2,
+          marginBottom: 2,
+          boxShadow: '0 2px 3px rgba(0,0,0,.1)',
+        }}
+        onClick={ hide } >
+        { isOpen && !isDisabled && children }
+      </Popper>
+    </Manager>
+  );
+});
 
 Dropdown.propTypes = {
+  /** Arrow icon direction:
+    * - Defaults to 'none', which hides the arrow
+    * - passing a prop value of false also hides the arrow
+    * - passing a prop value of true uses down arrow
+    */
+  arrowIcon: PropTypes.oneOf([
+    'down',
+    'left',
+    'none',
+    'right',
+    'up',
+  ]),
   /** Button text, can be a string or element. */
   buttonContent: PropTypes.oneOfType([
     PropTypes.string,
@@ -145,6 +134,10 @@ Dropdown.propTypes = {
   ]),
   /** Override default dropdown menu z-index. */
   zIndex: PropTypes.number,
+};
+
+Dropdown.defaultProps = {
+  arrowIcon: 'none',
 };
 
 export default Dropdown;