+++ src/components/Poptip/Poptip.story.js
@@ -3,6 +3,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withInfo } from '@storybook/addon-info';
+import { select } from '@storybook/addon-knobs';
+
 
 import Poptip from './index.js';
 import Button from '../Button';
@@ -22,6 +24,7 @@ stories
   .add('basic poptip', withInfo()(() => (<div className="position--relative height--100">
     <Poptip
       className="text--center"
+      trigger='mouseenter'
       content="Ipsa officiis bad-news minus earum a aperiam! Aperiam reiciendis vitae nihil libero et, hic!">
       <Button width="default">
         Basic Poptip
@@ -30,11 +33,9 @@ stories
   </div>)))
   .add('poptip to the right', withInfo()(() => (<div className="position--relative height--100">
     <Poptip
+      trigger='click'
       content="Ipsa officiis bad-news minus earum a aperiam! Aperiam reiciendis vitae nihil libero et, hic!"
-      horizontalAttachment="left"
-      horizontalTargetAttachment="right"
-      verticalAttachment="middle"
-      verticalTargetAttachment="middle">
+      theme='light'>
       <Button style="unstyled" width="default">
         <a>Poptip on the right</a>
       </Button>
@@ -42,11 +43,9 @@ stories
   </div>)))
   .add('poptip with advanced positioning', withInfo()(() => (<div className="position--relative height--100">
     <Poptip
-      content="Ipsa officiis bad-news minus earum a aperiam! Aperiam reiciendis vitae nihil libero et, hic!"
-      horizontalAttachment="left"
-      horizontalTargetAttachment="left"
-      verticalAttachment="top"
-      verticalTargetAttachment="bottom">
+      content={ <div>hola a todos mis amigos</div> }
+      trigger='click'
+      position={ select('position', {top: 'top', bottom: 'bottom', left: 'left', right: 'right'}, 'top') }>
       <Button width="default">
         Poptip With Advanced Positioning
       </Button>