+++ src/components/Toggle/Toggle.story.js
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import { storiesOf } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
+import { withKnobs } from '@storybook/addon-knobs';
+import { withInfo } from '@storybook/addon-info';
+
+import Toggle from './index.js';
+
+const stories = storiesOf('Toggle', module);
+stories
+  .addDecorator(withKnobs)
+  .addDecorator(story => (
+    <div id="root-preview">
+      {story()}
+    </div>
+  ));
+
+stories.add('default', withInfo()(() => {
+  return <Toggle onClick={ action('clicked on toggle') }/>;
+}));
+
+